requirements.txt
streamlit
pdf417gen
Pillow
pandas==2.1.4
pytz